moved net functions from ds3 to ds3_net
