+++ src/ds3.c
@@ -36,6 +36,7 @@
 #endif
 
 #include "ds3.h"
+#include "ds3_net.h"
 
 //---------- Define opaque struct ----------//
 struct _ds3_request{
@@ -302,18 +303,6 @@ static void _ds3_internal_str_free(gpointer data) {
     ds3_str_free((ds3_str*)data);
 }
 
-static void _ds3_free_response_header(gpointer data) {
-    ds3_response_header* header;
-    if (data == NULL) {
-        return;
-    }
-
-    header = (ds3_response_header*) data;
-    ds3_str_free(header->key);
-    g_ptr_array_free(header->values, TRUE);
-    g_free(data);
-}
-
 static ds3_str* _ds3_response_header_get_first(const ds3_response_header* header) {
     return (ds3_str*)g_ptr_array_index(header->values, 0);
 }
@@ -451,468 +440,11 @@ static size_t _process_response_body(void* buffer, size_t size, size_t nmemb, vo
     }
 }
 
-//---------- Networking code ----------//
-static void _init_curl(void) {
-    static ds3_bool initialized = False;
-
-    if (!initialized) {
-        if (curl_global_init(CURL_GLOBAL_ALL) != 0) {
-          fprintf(stderr, "Encountered an error initializing libcurl\n");
-        }
-        initialized = True;
-    }
-}
-
-static char* _net_get_verb(http_verb verb) {
-    switch(verb) {
-        case HTTP_GET: return "GET";
-        case HTTP_PUT: return "PUT";
-        case HTTP_POST: return "POST";
-        case HTTP_DELETE : return "DELETE";
-        case HTTP_HEAD : return "HEAD";
-    }
-
-    return NULL;
-}
-
-static char* _get_object_type(const object_type type) {
-    switch(type) {
-        case DATA: return "DATA";
-        case NO_TYPE: return NULL;
-    }
-
-    return NULL;
-}
-
-// curl_easy_escape'd strings must be freed using curl_free.  Copy
-// the escaped string, using glib, since users of this function will
-// eventually wind up freeing it with g_free.
-static char* _escape_url(const char* url) {
-    char* curl_escaped_url = curl_easy_escape(NULL, url, 0);
-    char* escaped_url = g_strdup(curl_escaped_url);
-    curl_free(curl_escaped_url);
-    return escaped_url;
-}
-
-// Like _escape_url but don't encode "/".
-static char* _escape_url_object_name(const char* url) {
-    gchar** split = g_strsplit(url, "/", 0);
-    gchar** ptr;
-    gchar* escaped_ptr;
-    for (ptr = split; *ptr; ptr++) {
-        escaped_ptr = _escape_url(*ptr);
-        g_free(*ptr);
-        *ptr = escaped_ptr;
-    }
-    escaped_ptr = g_strjoinv("/", split);
-    g_strfreev(split);
-    return escaped_ptr;
-}
-
-static unsigned char* _generate_signature_str(http_verb verb, char* resource_name, char* date,
-                               char* content_type, char* md5, char* amz_headers) {
-    char* verb_str;
-    if (resource_name == NULL) {
-        fprintf(stderr, "resource_name is required\n");
-        return NULL;
-    }
-    if (date == NULL) {
-        fprintf(stderr, "date is required");
-        return NULL;
-    }
-    verb_str = _net_get_verb(verb);
-
-    return (unsigned char*) g_strconcat(verb_str, "\n", md5, "\n", content_type, "\n", date, "\n", amz_headers, resource_name, NULL);
-}
-
-static char* _generate_date_string(void) {
-    GDateTime* time  = g_date_time_new_now_local();
-    char* date_string = g_date_time_format(time, "%a, %d %b %Y %H:%M:%S %z");
-
-    g_date_time_unref(time);
-
-    return date_string;
-}
-
-static char* _net_compute_signature(const ds3_log* log, const ds3_creds* creds, http_verb verb, char* resource_name,
-                             char* date, char* content_type, char* md5, char* amz_headers) {
-    GHmac* hmac;
-    gchar* signature;
-    gsize bufSize = 256;
-    guint8 buffer[256];
-
-    unsigned char* signature_str = _generate_signature_str(verb, resource_name, date, content_type, md5, amz_headers);
-    char* escaped_str = g_strescape((char*) signature_str, NULL);
-
-    LOG(log, DS3_DEBUG, "signature string: %s", escaped_str);
-    g_free(escaped_str);
-
-    hmac = g_hmac_new(G_CHECKSUM_SHA1, (unsigned char*) creds->secret_key->value, creds->secret_key->size);
-    g_hmac_update(hmac, signature_str, strlen((const char*)signature_str));
-    g_hmac_get_digest(hmac, buffer, &bufSize);
-
-    signature = g_base64_encode(buffer, bufSize);
-
-    g_free(signature_str);
-    g_hmac_unref(hmac);
-
-    return signature;
-}
-
-typedef struct {
-    char** entries;
-    size_t size;
-}query_entries;
-
-static void _hash_for_each(gpointer _key, gpointer _value, gpointer _user_data) {
-    char* key = (char*) _key;
-    char* value = (char*) _value;
-    query_entries* entries = (query_entries*) _user_data;
-    if (value == NULL) {
-        entries->entries[entries->size] = g_strconcat(key, NULL);
-    } else {
-        entries->entries[entries->size] = g_strconcat(key, "=", value, NULL);
-    }
-    entries->size++;
-}
-
-static char* _net_gen_query_params(GHashTable* query_params) {
-    if (g_hash_table_size(query_params) > 0) {
-        query_entries q_entries;
-        char** entries;
-        char* return_string;
-        int i;
-        //build the query string
-        memset(&q_entries, 0, sizeof(query_entries));
-
-        //We need the +1 so that it is NULL terminating for g_strjoinv
-        entries = g_new0(char*, g_hash_table_size(query_params) + 1);
-        q_entries.entries = entries;
-        g_hash_table_foreach(query_params, _hash_for_each, &q_entries);
-
-        return_string = g_strjoinv("&", entries);
-
-        for (i = 0; ; i++) {
-            char* current_string = entries[i];
-            if (current_string == NULL) {
-                break;
-            }
-            g_free(current_string);
-        }
-
-        g_free(entries);
-        return return_string;
-    } else {
-        return NULL;
-    }
-}
-
-static struct curl_slist* _append_headers(struct curl_slist* header_list, GHashTable* headers_map) {
-    GHashTableIter iter;
-    gpointer key, value;
-    struct curl_slist* updated_list = header_list;
-    g_hash_table_iter_init(&iter, headers_map);
-
-    while (g_hash_table_iter_next(&iter, &key, &value)) {
-        char* header_value = g_strconcat((char*)key, ": ", (char*)value, NULL);
-        updated_list = curl_slist_append(updated_list, header_value);
-        g_free(header_value);
-    }
-    return updated_list;
-}
-
-static int ds3_curl_logger(CURL *handle, curl_infotype type, char* data, size_t size, void* userp) {
-    char* text = "curl_log";
-    ds3_log* log = (ds3_log*) userp;
-    char* message;
-    switch(type) {
-        case CURLINFO_HEADER_OUT:
-          text = "HEADER_SENT";
-          break;
-        case CURLINFO_HEADER_IN:
-          text = "HEADER_RECV";
-          break;
-
-        case CURLINFO_DATA_IN:
-        case CURLINFO_DATA_OUT:
-        case CURLINFO_SSL_DATA_IN:
-        case CURLINFO_SSL_DATA_OUT:
-          // do not log any payload data
-          return 0;
-        default:
-          break;
-    }
-
-    message = g_strndup(data, size);
-
-    LOG(log, DS3_TRACE, "%s: %s", text, g_strchomp(message));
-
-    g_free(message);
-    return 0;
-}
-
-static gint _gstring_sort(gconstpointer a, gconstpointer b) {
-    char** val1 = (char**)a;
-    char** val2 = (char**)b;
-
-    return g_strcmp0(*val1, *val2);
-}
-
-static char* _canonicalize_amz_headers(GHashTable* headers) {
-    GList* keys = g_hash_table_get_keys(headers);
-    GList* key = keys;
-    GString* canonicalized_headers = g_string_new("");
-    GPtrArray *signing_strings = g_ptr_array_new_with_free_func(g_free);  // stores char*
-    GString* header_signing_value;
-    char* signing_value;
-    int i;
-
-    while(key != NULL) {
-        if(g_str_has_prefix((char*)key->data, "x-amz")){
-            header_signing_value = g_string_new((gchar*)key->data);
-            header_signing_value = g_string_append(g_string_ascii_down(header_signing_value), ":");
-            header_signing_value = g_string_append(header_signing_value, (gchar*)g_hash_table_lookup(headers, key->data));
-
-            signing_value = g_string_free(header_signing_value, FALSE);
-            g_ptr_array_add(signing_strings, signing_value);
-        }
-        key = key->next;
-    }
-
-    g_ptr_array_sort(signing_strings, _gstring_sort);
-
-    for (i = 0; i < signing_strings->len; i++) {
-        g_string_append(canonicalized_headers, (gchar*)g_ptr_array_index(signing_strings, i));
-        g_string_append(canonicalized_headers, "\n");
-    }
-
-    g_list_free(keys);
-    g_ptr_array_free(signing_strings, TRUE);
-
-    return g_string_free(canonicalized_headers, FALSE);
-}
-
-static char* _canonicalized_resource(ds3_str* path, GHashTable* query_params) {
-    if (g_hash_table_contains(query_params, "delete")) {
-        return g_strconcat(path->value, "?delete", NULL);
-    } else {
-        return g_strdup(path->value);
-    }
-}
-
-static ds3_error* _net_process_request(const ds3_client* client, const ds3_request* _request, void* read_user_struct, size_t (*read_handler_func)(void*, size_t, size_t, void*), void* write_user_struct, size_t (*write_handler_func)(void*, size_t, size_t, void*), GHashTable** return_headers) {
-    struct _ds3_request* request = (struct _ds3_request*) _request;
-    CURL* handle;
-    CURLcode res;
-    char* url;
-    int retry_count = 0;
-    char* query_params;
-
-    _init_curl();
-
-    query_params = _net_gen_query_params(request->query_params);
-
-    if (query_params == NULL) {
-        url = g_strconcat(client->endpoint->value, request->path->value, NULL);
-    } else {
-        url = g_strconcat(client->endpoint->value, request->path->value,"?",query_params, NULL);
-        g_free(query_params);
-    }
-
-    while (retry_count < client->num_redirects) {
-        handle = curl_easy_init();
-
-        if (handle) {
-            char* amz_headers;
-            char* canonicalized_resource;
-            char* date;
-            char* date_header;
-            char* signature;
-            struct curl_slist* headers;
-            char* auth_header;
-            char* md5_value;
-            ds3_response_data response_data;
-            GHashTable* response_headers = g_hash_table_new_full(g_str_hash, g_str_equal, g_free, _ds3_free_response_header);
-
-            LOG(client->log, DS3_DEBUG, "Preparing to send request");
-
-            memset(&response_data, 0, sizeof(ds3_response_data));
-            response_data.headers = response_headers;
-            response_data.body = g_byte_array_new();
-
-            if (client->log != NULL) {
-                curl_easy_setopt(handle, CURLOPT_DEBUGFUNCTION, ds3_curl_logger);
-                curl_easy_setopt(handle, CURLOPT_DEBUGDATA, client->log);
-                curl_easy_setopt(handle, CURLOPT_VERBOSE, 1L); // turn on verbose logging
-            }
-
-            curl_easy_setopt(handle, CURLOPT_URL, url);
-
-            curl_easy_setopt(handle, CURLOPT_FOLLOWLOCATION, 0); // explicitly disable
-
-            // Setup header collection
-            curl_easy_setopt(handle, CURLOPT_HEADERFUNCTION, _process_header_line);
-            curl_easy_setopt(handle, CURLOPT_HEADERDATA, &response_data);
-
-            if (client->proxy != NULL) {
-              curl_easy_setopt(handle, CURLOPT_PROXY, client->proxy->value);
-            }
-
-            // Register the read and write handlers if they are set
-            if (read_user_struct != NULL && read_handler_func != NULL) {
-                response_data.user_data = read_user_struct;
-                response_data.user_func = read_handler_func;
-            }
-
-            // We must always set this so we can collect the error message body
-            curl_easy_setopt(handle, CURLOPT_WRITEFUNCTION, _process_response_body);
-            curl_easy_setopt(handle, CURLOPT_WRITEDATA, &response_data);
-
-            if (write_user_struct != NULL && write_handler_func != NULL) {
-                curl_easy_setopt(handle, CURLOPT_READFUNCTION, write_handler_func);
-                curl_easy_setopt(handle, CURLOPT_READDATA, write_user_struct);
-            }
-
-            switch(request->verb) {
-                case HTTP_POST: {
-                    curl_easy_setopt(handle, CURLOPT_CUSTOMREQUEST, "POST");
-                    curl_easy_setopt(handle, CURLOPT_UPLOAD, 1L);
-                    curl_easy_setopt(handle, CURLOPT_INFILESIZE_LARGE, request->length);
-                    break;
-                }
-                case HTTP_PUT: {
-                    curl_easy_setopt(handle, CURLOPT_PUT, 1L);
-                    curl_easy_setopt(handle, CURLOPT_UPLOAD, 1L);
-                    curl_easy_setopt(handle, CURLOPT_INFILESIZE_LARGE, request->length);
-                    break;
-                }
-                case HTTP_DELETE: {
-                    curl_easy_setopt(handle, CURLOPT_CUSTOMREQUEST, "DELETE");
-                    break;
-                }
-                case HTTP_HEAD: {
-                    curl_easy_setopt(handle, CURLOPT_CUSTOMREQUEST, "HEAD");
-                    curl_easy_setopt(handle, CURLOPT_NOBODY, 1L);
-                    break;
-                }
-                case HTTP_GET: {
-                    //Placeholder if we need to put anything here.
-                    break;
-                }
-            }
-
-            date = _generate_date_string();
-            date_header = g_strconcat("Date: ", date, NULL);
-            headers = NULL;
-
-            if (request->md5 == NULL) {
-                md5_value = "";
-            } else {
-                char* md5_header;
-                md5_value = request->md5->value;
-                md5_header = g_strconcat("Content-MD5:", md5_value, NULL);
-                headers = curl_slist_append(headers, md5_header);
-                g_free(md5_header);
-            }
-            amz_headers = _canonicalize_amz_headers(request->headers);
-            canonicalized_resource = _canonicalized_resource(request->path, request->query_params);
-            signature = _net_compute_signature(client->log, client->creds, request->verb, canonicalized_resource, date, "", md5_value, amz_headers);
-
-            g_free(amz_headers);
-            g_free(canonicalized_resource);
-
-            auth_header = g_strconcat("Authorization: AWS ", client->creds->access_id->value, ":", signature, NULL);
-
-            headers = curl_slist_append(headers, auth_header);
-            headers = curl_slist_append(headers, date_header);
-            headers = _append_headers(headers, request->headers);
-
-            curl_easy_setopt(handle, CURLOPT_HTTPHEADER, headers);
-
-            res = curl_easy_perform(handle);
-
-            g_free(date);
-            g_free(date_header);
-            g_free(signature);
-            g_free(auth_header);
-            curl_slist_free_all(headers);
-            curl_easy_cleanup(handle);
-
-            //process the response
-            if (res != CURLE_OK) {
-                char * message = g_strconcat("Request failed: ", curl_easy_strerror(res), NULL);
-                ds3_error* error = _ds3_create_error(DS3_ERROR_REQUEST_FAILED, message);
-                g_free(url);
-                g_byte_array_free(response_data.body, TRUE);
-                ds3_str_free(response_data.status_message);
-                g_hash_table_destroy(response_headers);
-                g_free(message);
-                return error;
-            }
-
-            LOG(client->log, DS3_DEBUG, "Request completed with status code of: %d", response_data.status_code);
-
-            if (response_data.status_code == 307) {
-                LOG(client->log, DS3_INFO, "Request encountered a 307 redirect");
-                ds3_str_free(response_data.status_message);
-
-                if (response_data.body != NULL) {
-                    g_byte_array_free(response_data.body, TRUE);
-                }
-                g_hash_table_destroy(response_headers);
-                retry_count++;
-                LOG(client->log, DS3_DEBUG, "Retry Attempt: %d | Max Retries: %d", retry_count, client->num_redirects);
-                continue;
-            }
-
-            if (response_data.status_code < 200 || response_data.status_code >= 300) {
-                ds3_error* error = _ds3_create_error(DS3_ERROR_BAD_STATUS_CODE, "Got an unexpected status code.");
-                error->error = g_new0(ds3_error_response, 1);
-                error->error->status_code = response_data.status_code;
-                error->error->status_message = ds3_str_init(response_data.status_message->value);
-                if (response_data.body != NULL) {
-                    error->error->error_body = ds3_str_init_with_size((char*)response_data.body->data, response_data.body->len);
-                    g_byte_array_free(response_data.body, TRUE);
-                } else {
-                    LOG(client->log, DS3_ERROR, "The response body for the error is empty");
-                    error->error->error_body = NULL;
-                }
-                g_hash_table_destroy(response_headers);
-                ds3_str_free(response_data.status_message);
-                g_free(url);
-                return error;
-            }
-
-            g_byte_array_free(response_data.body, TRUE);
-            ds3_str_free(response_data.status_message);
-            if (return_headers == NULL) {
-                g_hash_table_destroy(response_headers);
-            } else {
-                *return_headers = response_headers;
-            }
-
-            break;
-        } else {
-            return _ds3_create_error(DS3_ERROR_CURL_HANDLE, "Failed to create curl handle");
-        }
-    }
-    g_free(url);
-
-    if (retry_count == client->num_redirects) {
-      return _ds3_create_error(DS3_ERROR_TOO_MANY_REDIRECTS, "Encountered too many redirects while attempting to fullfil the request");
-    }
-    return NULL;
-}
-
-static void net_cleanup(void) {
-    curl_global_cleanup();
-}
-
+//---------- Ds3 code ----------//
 static void _cleanup_hash_value(gpointer value) {
     g_free(value);
 }
 
-//---------- Ds3 code ----------//
 static GHashTable* _create_hash_table(void) {
     GHashTable* hash =  g_hash_table_new_full(g_str_hash, g_str_equal, _cleanup_hash_value, _cleanup_hash_value);
     return hash;
@@ -984,12 +516,12 @@ ds3_error* ds3_create_client_from_env(ds3_client** client) {
 }
 
 static void _set_map_value(GHashTable* map, const char* key, const char* value) {
-    gpointer escaped_key = (gpointer) _escape_url(key);
+    gpointer escaped_key = (gpointer) escape_url(key);
 
     //TODO update this to handle multiple values being set for a header field
     gpointer escaped_value;
     if (value != NULL) {
-        escaped_value = (gpointer) _escape_url(value);
+        escaped_value = (gpointer) escape_url(value);
     } else {
         escaped_value = NULL;
     }
@@ -1055,6 +587,15 @@ void ds3_request_set_id(ds3_request* _request, const char* id) {
     _set_query_param(_request, "id", id);
 }
 
+static char* _get_object_type(const object_type type) {
+    switch(type) {
+        case DATA: return "DATA";
+        case NO_TYPE: return NULL;
+    }
+
+    return NULL;
+}
+
 void ds3_request_set_type(ds3_request* _request, object_type type) {
     const char* type_as_string = _get_object_type(type);
     if(type_as_string != NULL) {
@@ -1083,10 +624,10 @@ static ds3_str* _build_path(const char* path_prefix, const char* bucket_name, co
     char* full_path = NULL;
 
     if (bucket_name != NULL) {
-        escaped_bucket_name = _escape_url(bucket_name);
+        escaped_bucket_name = escape_url(bucket_name);
     }
     if (object_name != NULL) {
-        escaped_object_name = _escape_url_object_name(object_name);
+        escaped_object_name = escape_url_object_name(object_name);
     }
 
     joined_path = g_strjoin("/", escaped_bucket_name, escaped_object_name, NULL);
@@ -1284,7 +825,7 @@ static ds3_error* _internal_request_dispatcher(const ds3_client* client, const d
     if (client == NULL || request == NULL) {
         return _ds3_create_error(DS3_ERROR_MISSING_ARGS, "All arguments must be filled in for request processing");
     }
-    return _net_process_request(client, request, read_user_struct, read_handler_func, write_user_struct, write_handler_func, NULL);
+    return net_process_request(client, request, read_user_struct, read_handler_func, write_user_struct, write_handler_func, NULL);
 }
 
 static bool attribute_equal(const struct _xmlAttr* attribute, const char* attribute_name){
@@ -1803,7 +1344,7 @@ ds3_error* ds3_head_object(const ds3_client* client, const ds3_request* request,
         return _ds3_create_error(DS3_ERROR_MISSING_ARGS, "The object name parameter is required.");
     }
 
-    error = _net_process_request(client, request, NULL, NULL, NULL, NULL, &return_headers);
+    error = net_process_request(client, request, NULL, NULL, NULL, NULL, &return_headers);
 
     if (error == NULL) {
         fprintf(stderr, "Head object completed successfully\n");
@@ -1818,7 +1359,7 @@ ds3_error* ds3_head_object(const ds3_client* client, const ds3_request* request,
 ds3_error* ds3_head_bucket(const ds3_client* client, const ds3_request* request) {
     ds3_error* error;
 
-    error = _net_process_request(client, request, NULL, NULL, NULL, NULL, NULL);
+    error = net_process_request(client, request, NULL, NULL, NULL, NULL, NULL);
 
     return error;
 }
@@ -1832,7 +1373,7 @@ ds3_error* ds3_get_object_with_metadata(const ds3_client* client, const ds3_requ
     GHashTable* return_headers;
     ds3_metadata* metadata;
 
-    error = _net_process_request(client, request, user_data, callback, NULL, NULL, &return_headers);
+    error = net_process_request(client, request, user_data, callback, NULL, NULL, &return_headers);
 
     if (error == NULL) {
         fprintf(stderr, "Head object completed successfully\n");
@@ -2331,7 +1872,7 @@ ds3_error* ds3_delete_objects(const ds3_client* client, const ds3_request* _requ
 
     xml_blob = g_byte_array_new();
 
-    error_response = _net_process_request(client, request, xml_blob, load_buffer, (void*) &send_buff, _ds3_send_xml_buff, NULL);
+    error_response = net_process_request(client, request, xml_blob, load_buffer, (void*) &send_buff, _ds3_send_xml_buff, NULL);
 
     // Cleanup the data sent to the server.
     xmlFreeDoc(doc);
@@ -2344,7 +1885,7 @@ ds3_error* ds3_delete_objects(const ds3_client* client, const ds3_request* _requ
 ds3_error* ds3_delete_folder(const ds3_client* client, const ds3_request* _request) {
     struct _ds3_request* request;
     request = (struct _ds3_request*) _request;
-    return _net_process_request(client, request, NULL, NULL, NULL, NULL, NULL);
+    return net_process_request(client, request, NULL, NULL, NULL, NULL, NULL);
 }
 
 ds3_error* ds3_get_physical_placement(const ds3_client* client, const ds3_request* _request, ds3_get_physical_placement_response** _response){
@@ -2394,7 +1935,7 @@ ds3_error* ds3_get_physical_placement(const ds3_client* client, const ds3_reques
     request->length = send_buff.size; // make sure to set the size of the request.
 
     xml_blob = g_byte_array_new();
-    error_response = _net_process_request(client, request, xml_blob, load_buffer, (void*) &send_buff, _ds3_send_xml_buff, NULL);
+    error_response = net_process_request(client, request, xml_blob, load_buffer, (void*) &send_buff, _ds3_send_xml_buff, NULL);
 
     // Cleanup the data sent to the server.
     xmlFreeDoc(doc);
@@ -2557,7 +2098,7 @@ ds3_error* ds3_bulk(const ds3_client* client, const ds3_request* _request, ds3_b
     request->length = send_buff.size; // make sure to set the size of the request.
 
     xml_blob = g_byte_array_new();
-    error_response = _net_process_request(client, request, xml_blob, load_buffer, (void*) &send_buff, _ds3_send_xml_buff, NULL);
+    error_response = net_process_request(client, request, xml_blob, load_buffer, (void*) &send_buff, _ds3_send_xml_buff, NULL);
 
     // Cleanup the data sent to the server.
     xmlFreeDoc(doc);
@@ -2599,7 +2140,7 @@ ds3_error* ds3_allocate_chunk(const ds3_client* client, const ds3_request* reque
     xmlDocPtr doc;
     xmlNodePtr root;
 
-    error = _net_process_request(client, request, xml_blob, load_buffer, NULL, NULL, &response_headers);
+    error = net_process_request(client, request, xml_blob, load_buffer, NULL, NULL, &response_headers);
 
     if (error != NULL) {
         g_byte_array_free(xml_blob, TRUE);
@@ -2656,7 +2197,7 @@ ds3_error* ds3_get_available_chunks(const ds3_client* client, const ds3_request*
     ds3_response_header* retry_after_header;
     xmlDocPtr doc;
 
-    error = _net_process_request(client, request, xml_blob, load_buffer, NULL, NULL, &response_headers);
+    error = net_process_request(client, request, xml_blob, load_buffer, NULL, NULL, &response_headers);
 
     if (error != NULL) {
         if (response_headers != NULL) {
@@ -2743,7 +2284,7 @@ ds3_error* ds3_get_jobs(const ds3_client* client, const ds3_request* request, ds
     GHashTable* response_headers = NULL;
     xmlDocPtr doc;
 
-    error = _net_process_request(client, request, xml_blob, load_buffer, NULL, NULL, &response_headers);
+    error = net_process_request(client, request, xml_blob, load_buffer, NULL, NULL, &response_headers);
     if (error != NULL) {
         g_byte_array_free(xml_blob, TRUE);
         return error;
@@ -2772,7 +2313,7 @@ static ds3_error* _common_job(const ds3_client* client, const ds3_request* reque
     ds3_bulk_response* bulk_response;
     xmlDocPtr doc;
 
-    error = _net_process_request(client, request, xml_blob, load_buffer, NULL, NULL, NULL);
+    error = net_process_request(client, request, xml_blob, load_buffer, NULL, NULL, NULL);
 
     if (error != NULL) {
         g_byte_array_free(xml_blob, TRUE);
